// ui/src/renderer/src/houston/generated/UpdateState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UpdateRelease } from "./UpdateRelease";

export type UpdateState = { "kind": "unknown" } | { "kind": "disabled" } | { "kind": "checking" } | { "kind": "up_to_date", checked_at_ms: number, } | { "kind": "available", release: UpdateRelease, checked_at_ms: number, } | { "kind": "failed", error: string, checked_at_ms: number, };